rename menu component folder and move MenuList file in
