+++ src/components/menu/MenuList.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import MenuItem from "../menu-item/MenuItem";
+import MenuItem from "./MenuItem";
 import Sponsor from "../sponsor/Sponsor";
 
 const VALUES = {
